Sff8472 can't support both SFP/QSFP so breaking it part per type
